feat(filter): 将分页参数类型从 int 改为 int64
修改了 Filter 接口及其实现中的 Pagination 方法签名，将 pageNum 和 pageSize
的类型由 int 改为 int64，以支持更大的分页数值范围。同时更新了相关测试代码和
工具类中对分页结构体字段类型的定义，并修正了一个常量命名错误。

+++ models/filter.go
@@ -44,7 +44,7 @@ type Filter interface {
 	In(key string, val any) *cd.Error
 	NotIn(key string, val any) *cd.Error
 	Like(key string, val any) *cd.Error
-	Pagination(pageNum, pageSize int)
+	Pagination(pageNum, pageSize int64)
 	Sort(fieldName string, ascFlag bool)
 	ValueMask(val any) *cd.Error
